Introduce utils/osmo-amr-inspect program
This program allows inspecting a list of AMR PDUs in hexstring format
present in a text file.
This allows easily finding out the properties of the PDU as well as
finding potential errors on the data.
It also allows forcing decode in different formats (octet-aligned,
bandwidth-efficient) in order to detect which may be the correct one.

Related: SYS#6161
Change-Id: Iffa6cc2e5391b77e3097d4c3b8d3f5211427dbe2