fix msgb leak on OSMO_STREAM_SRV_F_FLUSH_DESTROY
In osmo_stream_srv_send(), call msgb_free() if the msgb is not enqueued.

Before this, osmo_stream_srv_send() enqueues a msgb, assuming the queue
will free it. But if OSMO_STREAM_SRV_F_FLUSH_DESTROY is set, it just
returns and drops the msgb. There is no rc, so the caller cannot know.

Change-Id: I994c5e06a9030cfbf4fca8ab42588c61f7121cee