stream_cli.c: Handle read / recvfrom error and close connection
If read or recvfrom fails or returns 0, the connection must be closed.
This is already done when a write / send fails. In both cases the
disconnect callback is called to notify the user's client.

Not handling the error may cause an infinite loop of read or recvfrom
failures.

Related: OS#6405
Change-Id: I55426de6b49cb4cb0797e50dfeae11f2efc29b15