stream_cli: Explicitly ignore return code of internal functions
Make coverity happy.
We don't really care about the return code of the stream_cli_reconnect()
function in the code path, since the return value is only used
internally/privately in the object and there's no further access/use of
the object after calling it in osmo_stream_cli_reconnect().

Change-Id: I61b7001047e8eb1e1347f532d67f3f00dc8f67ec