stream: Support configuring TCP keep-alive pars
for servers, the param can either be set on the link (which will be
automatically inherited/applied for all accepted conn sockets), or can
be set specifically on a srv conn socket.

The second is useful when user can supply different parameters for different
clients connected to the listening socket, eg. ASP configuration in
libosmo-sigtran.

Change-Id: Ie748ad581c1c42f4c24f9409ce7d34d419cbca8b